break up distributer a bit
